lmkd: consider lmkd upgraded events history for supercritical event
Consider lmkd upgraded events history for supercritical event. This is
because under high file cache levels in the system all the upgraded
events in the file_cache_to_adj() will fail to adjust the selection of
the adj of the process to kill thus pressure can build up in the system.
For some clients, this is manifested into memory allocation timeout
which can be avoided by timely kill.

CRs-Fixed: 2720759
Change-Id: I8ffd47a240cb4b13d3d53687e562a6ac542f3581